+++ packages/core/CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - When a user logs in, the `CartSessionAuthListener` will now check for an active cart, rather than just grabbing the latest. ([#186](https://github.com/getcandy/getcandy/issues/186))
+- `Dropdown`, `ListField` and `Number` FieldTypes now implement the `JsonSerializable` interface.
 
 ## 2.0-beta10 - 2022-02-18